added movieDAOs, add category to movie

+++ src/movieplayer/gui/controller/CategoryWindowController.java
@@ -2,6 +2,7 @@ package movieplayer.gui.controller;
 
 import java.io.IOException;
 import java.net.URL;
+import java.sql.SQLException;
 import java.util.ResourceBundle;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -14,6 +15,7 @@ import javafx.scene.control.ListView;
 import javafx.stage.Stage;
 import movieplayer.be.Category;
 import movieplayer.gui.model.CategoryModel;
+import movieplayer.gui.util.MessageBoxHelper;
 
 /**
  * FXML Controller class
@@ -32,8 +34,7 @@ public class CategoryWindowController implements Initializable {
      */
     @Override
     public void initialize(URL url, ResourceBundle rb) {
-        listviewCategories.setItems(cmodel.getCategories());
-        cmodel.getCategories();
+        reload();
     }
     
     @FXML
@@ -42,7 +43,9 @@ public class CategoryWindowController implements Initializable {
         Scene scene = new Scene(root);
         Stage stage = (Stage) new Stage();
         stage.setScene(scene);
-        stage.show();
+        stage.showAndWait();
+        
+        reload();
     }
     
     @FXML
@@ -52,8 +55,23 @@ public class CategoryWindowController implements Initializable {
     }
     
     @FXML private void deleteCategory(ActionEvent event) {
-        Category clickedCategory = listviewCategories.getSelectionModel().getSelectedItem();
-        cmodel.deleteCategory(clickedCategory);
+        try {
+            Category clickedCategory = listviewCategories.getSelectionModel().getSelectedItem();
+            if(clickedCategory != null) {
+                if(!MessageBoxHelper.askYesNo("Are you sure you want to delete the category " + clickedCategory.getName() + "?")) {
+                    return;
+                }
+                cmodel.deleteCategory(clickedCategory);
+                reload();
+            } else {
+                MessageBoxHelper.displayError("You must select a category.");
+            } 
+        } catch(SQLException ex) {
+            MessageBoxHelper.displayException(ex);
+        }
     }
     
+    private void reload() {
+        listviewCategories.setItems(cmodel.getCategories());
+    }
 }
